Create QueensAttack
Problem was to count number of moves available to the Queen on an arbitrary chessboard of nxn squares, with k obstacles.  My solution simply moves the queen until either an edge or an obstacle is found, counting moves as it goes.  Most other solutions find the closest obstacle to the queen in the 8 available directions and then compute number of free squares up to that obstacle or the edge, I am including this here as an alternative solution.